Extract GPU data structs into TreeCullingTypes.h
Move TreeCullData, TreeRenderDataGPU, WorldLeafInstanceGPU, VisibleTreeData,
CellCullParams, TreeFilterParams, LeafCullP3Params, TreeCellCullUniforms,
and TreeFilterUniforms to a shared header. This enables the upcoming stage
extraction to include these types independently.
